Refactor login command handler to follow the guidelines of the architecture - command handlers deal with WHAT needs to happen, not HOW
