+++ examples/bri-3/src/bri/auth/capabilities/login/login.command.handler.ts
@@ -1,7 +1,6 @@
 import { ICommandHandler, CommandHandler } from '@nestjs/cqrs';
 import { LoginCommand } from './login.command';
 import { AuthAgent } from '../../agent/auth.agent';
-import { errorMessage, jwtConstants } from '../../constants';
 
 @CommandHandler(LoginCommand)
 export class LoginCommandHandler implements ICommandHandler<LoginCommand> {
@@ -11,30 +10,11 @@ export class LoginCommandHandler implements ICommandHandler<LoginCommand> {
     const { message, signature, publicKey } = command;
 
     const bpiSubject = await this.authAgent.getBpiSubjectByPublicKey(publicKey);
-    if (bpiSubject.loginNonce !== message) {
-      throw new Error(errorMessage.USER_NOT_AUTHORIZED);
-    }
 
-    const verifySignature = this.authAgent.verify(
-      message,
-      signature,
-      publicKey,
-    );
-    if (!verifySignature) {
-      throw new Error(errorMessage.USER_NOT_AUTHORIZED);
-    }
+    this.authAgent.throwIfLoginNonceMismatch(bpiSubject, message);
 
-    const serviceUrl = process.env.SERVICE_URL;
-    const subjectDid = `did:ethr:0x5:${publicKey}`;
-    const now = Math.floor(Date.now() / 1000);
-    const didJwtPayload = {
-      aud: serviceUrl,
-      sub: subjectDid,
-      exp: `${now + Math.floor(jwtConstants.expiresIn / 1000)}`,
-      nbf: `${now}`,
-      iat: `${now}`,
-    };
+    this.authAgent.throwIfSignatureVerificationFails(message, signature, publicKey);
 
-    return this.authAgent.generateDidJwt(didJwtPayload);
+    return this.authAgent.generateDidJwt(bpiSubject);
   }
 }
